+++ layers/Engine/packages/component-model/src/DirectiveResolvers/ResolveValueAndMergeFieldDirectiveResolver.php
@@ -435,7 +435,9 @@ final class ResolveValueAndMergeFieldDirectiveResolver extends AbstractGlobalFie
         );
 
         // 3. Add the output in the DB
-        if ($objectTypeFieldResolutionFeedbackStore->getErrors() !== []) {
+        if ($objectTypeFieldResolutionFeedbackStore->getErrors() !== []
+            || $objectTypeFieldResolutionFeedbackStore->getPartialErrors() !== []
+        ) {
             // Set the response for the failing field as null
             $resolvedIDFieldValues[$id][$field] = null;
             $this->setAppStateForFieldValuePromises(