+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -18,6 +18,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Added new input `IncludeExcludeFilterInput` (#3127)
 - Added documentation for new fields `_strArrayReplace` and `_strArrayReplaceMultiple` from the **PHP Functions via Schema** extension (#3129)
 - Added field `optionObjectValues` (#3133)
+- Support for "partial" errors (shown in the response, but not bubbled up on nested directives) (#3135)
 
 ### Improvements